// generated/src/app/api-models/cap-trans-ytd.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CapTransYtd {
  seqCapTransYtd: number;
  capModelId: string;
  firstContractMonth: Date;
  capEntityCode: string;
  seqCapPoolId: number;
  seqProvId: number;
  seqVendId: number;
  transType: string;
  capStoplossId: string;
  capTransYtdAmt: number;
  pmpmWitholdAmt: number;
}